requirements.txt
google-cloud-vision
google-auth
google-auth-oauthlib
google-auth-httplib2
